Feature/auth

+++ modules/application/src/main/java/edu/northeastern/truthtree/controller/stories/Stories.java
@@ -1,6 +1,16 @@
 package edu.northeastern.truthtree.controller.stories;
 
+import edu.northeastern.truthtree.dto.NameValueDTO;
+import edu.northeastern.truthtree.dto.StoryDTO;
 import edu.northeastern.truthtree.dto.StoryPaginationResponseDTO;
+import edu.northeastern.truthtree.enums.OrderType;
+import edu.northeastern.truthtree.enums.StoryStatus;
+import edu.northeastern.truthtree.enums.VoteType;
+import edu.northeastern.truthtree.service.stories.IStoriesService;
+import java.util.List;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -9,15 +19,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
-
-import edu.northeastern.truthtree.dto.NameValueDTO;
-import edu.northeastern.truthtree.dto.StoryDTO;
-import edu.northeastern.truthtree.enums.OrderType;
-import edu.northeastern.truthtree.enums.StoryStatus;
-import edu.northeastern.truthtree.enums.VoteType;
-import edu.northeastern.truthtree.service.stories.IStoriesService;
 
 @CrossOrigin(origins = "*", maxAge = 3600)
 @RestController
@@ -40,14 +41,29 @@ public class Stories implements IStories {
       @RequestParam(value = "orderType", required = false) OrderType orderType,
       @RequestParam(value = "storyStatus", required = false) StoryStatus storyStatus,
       @RequestParam(value = "pageSize", required = false) Integer pageSize,
-      @RequestParam(value = "currentPage", required = false) Integer currentPage
+      @RequestParam(value = "currentPage", required = false) Integer currentPage,
+      HttpServletRequest httpServletRequest, HttpServletResponse httpServletResponse
   ) {
+    if (storyStatus != null && !storyStatus.equals(StoryStatus.APPROVED)) {
+      HttpSession httpSession = httpServletRequest.getSession();
+      if (httpSession.getAttribute("admin") == null) {
+        httpServletResponse.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
+        return null;
+      }
+
+    }
     return service.getStories(orderType, storyStatus, pageSize, currentPage);
   }
 
   @RequestMapping(value = "/api/stories/story/{status}/{id}", method = RequestMethod.PUT)
-  public void changeStatus(@PathVariable StoryStatus status, @PathVariable String id) {
-    service.changeStatus(status, id);
+  public void changeStatus(@PathVariable StoryStatus status, @PathVariable String id,
+      HttpServletRequest httpServletRequest, HttpServletResponse httpServletResponse) {
+    HttpSession httpSession = httpServletRequest.getSession();
+    if (httpSession.getAttribute("admin") == null) {
+      httpServletResponse.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
+    } else {
+      service.changeStatus(status, id);
+    }
   }
 
   @RequestMapping(value = "/api/stories", method = RequestMethod.PUT)
@@ -58,8 +74,14 @@ public class Stories implements IStories {
 
   @Override
   @RequestMapping(value = "/api/stories/{id}", method = RequestMethod.DELETE)
-  public void deleteStory(@PathVariable String id) {
-    service.deleteStory(id);
+  public void deleteStory(@PathVariable String id, HttpServletRequest httpServletRequest,
+      HttpServletResponse httpServletResponse) {
+    HttpSession httpSession = httpServletRequest.getSession();
+    if (httpSession.getAttribute("admin") == null) {
+      httpServletResponse.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
+    } else {
+      service.deleteStory(id);
+    }
   }
 
   @RequestMapping(value = "/api/stories/order", method = RequestMethod.GET)
